Added translator for File class, made running basic SD example which reads from file into buffer, it's needed to implement Serial.write() into main transcriptor. Simple SD card functions are running ok.
As written above, need to implement function for serial line writing bytes, because for now I can write strings, but from files I'm reading array of bytes.
If I want to save array of bytes from file into vector, I have to provide pointer to first item. This is done by expression &vectorName.front() I found on internet that this doesn't have torun when array is empty, instead they recommended usage of function vectorName.data(), but that one is not running for me. THIS WITH ADDRESS OF FRONT is for now OK for ME :)