Added behavior to the password locked VIs, so now it's able to simulate it on desktop PC. Still need to see if listfiles.vi & listfiles2.vi behaves same way, I remember listfiles.vi is able to compile and run OK, listfiles2.vi is now behave good in simulation, need to try if it's able to compile and run  correctly.
I spend whole night playing with recursive traversing folders because I wanted to make SD card function openNextFile.vi run. Now it's running in way, that it will open next file or folder. If as input is folder this function traverse it down until there is no directory and there continue reading files.
So fo file it enters it, I had to put feedback nodes into functions to compare input parameter, if it's same as before it means that this is somewhere creating cycles. This is RUNNING! (now I have on runnable version for HW and one for simulation, hopefully itll not be hard to glue them together :D)