"SWITCH outputs IN "Translator SD Card Library.vi" FOR openNextFile.vi i DONT'T KNOW WHY, BUT RIGHT TOP CORNER HAS INDEX 1 AND TERMINAL UNDER IT HAS INDEX 0, also added evaluation if file was open, so there are two lines in translator.
I really don't knowbut it's so, that top rigth corner for openNextFile.vi is 1 and terminal under that is 1, maybe it will be needed to make it back next time, need to test on another computer.
Also from generated code there are listed just top directories, it's not going deeper.

If you include <ArduinoSTL.h> and <vector> it's not working properly, but these libraries are for me system libs, so this app is developed with them, probably it's internal conflict!

package was rebuilded.