Corrected open.vi(), there was error when file hasn't exist, file for getting some info was throwing error, now repaired, when file is open for writing it's not checking if it's dir or file, just go straight for file option, and if file not exist it will creates it (open or create is set), LIBRARY REBUILDED.
I corrected this problem so this library is closer to be fully operational in same way on desktop PC as on arduino. There could be some small bugs, but until now for me it looks OK.
I found this bug while creating DEMO to write samples of sinewave signal to file on SD card, because on arduino it will create file hopefully.
In my case when creating DEMO arduino freezes, because low memory, keep in mind that writing big chunks of data into file can be memory consuming and reset/freeze/else things your processor (embedded processor HAS LOW MEMORY, KEEP THIS IN MIND).